ADDING NEWLINE TO COMPLY WITH TCP FORWARDER
This '\n' is added to make sure the packets are correctly forwarded
to Darwin with all TCP forwarder.
Some TCP forwarder are holding on the TCP packet unless
the packet contains a newline '\n'.

This behavior was observed with :
  - HAProxy 2.2.5

This addition does not interfere with the json parsing made by Darwin.